+++ packages/common/src/Base/createBaseInsertPluginButton.jsx
@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 import { AtomicBlockUtils, EditorState, SelectionState } from '@wix/draft-js';
 import cloneDeep from 'lodash/cloneDeep';
@@ -13,7 +13,7 @@ import styles from '../../statics/styles/toolbar-button.scss';
  * createBaseInsertPluginButton
  */
 export default ({ blockType, button, helpers, pubsub, settings, t }) => {
-  class InsertPluginButton extends Component {
+  class InsertPluginButton extends React.PureComponent {
     constructor(props) {
       super(props);
       this.state = {};