+++ packages/plugin-line-spacing/src/types.js
@@ -1,0 +1,1 @@
+export const LINE_SPACING_TYPE = 'line-spacing';