+++ scripts/linkExample.js
@@ -19,10 +19,20 @@ const getRootDir = dirname => {
   }
 };
 
+const getYarnLinkDir = () => {
+  if (process.env.YARN_LINK_DIR) {
+    return path.resolve(process.env.YARN_LINK_DIR);
+  } else if (process.platform === 'win32' && process.env.LOCALAPPDATA) {
+    return path.resolve(process.env.LOCALAPPDATA, 'Yarn', 'config', 'link');
+  } else {
+    return path.resolve(process.env.HOME, '.config', 'yarn', 'link');
+  }
+};
+
 const rootDir = getRootDir(__dirname);
 const PATHS = {
   lernaJson: path.resolve(rootDir, 'lerna.json'),
-  yarnLink: path.resolve(process.env.HOME, '.config', 'yarn', 'link'),
+  yarnLink: getYarnLinkDir(),
   example: path.resolve(rootDir, 'examples', exampleName),
 };
 const packages = [];